Refactor: reorganise modules, relocate utility functions, rename modules, update import statements and unit tests
