+++ src/wf_psf/data/data_handler.py
@@ -18,6 +18,9 @@ import wf_psf.utils.utils as utils
 from wf_psf.psf_models.tf_modules.tf_utils import ensure_tensor
 import tensorflow as tf
 from typing import Optional, Union
+import numpy as np
+import tensorflow as tf
+from fractions import Fraction
 import logging
 
 logger = logging.getLogger(__name__)
